Decorator Pattern of Design pattern
Decorator Pattern extends the function of class via decorators who implement original class or interface.
This pattern do only to extend and it don't care of what is the decorated class doing.

For using decorator pattern, We need abstract class that will be extended by original class and original class and abstract class that will be extended by decorator classes and decorator class.

Decorator Pattern was often used to classes of java.io.